refactor: fix service layer tests
